backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
gunicorn
django-cors-headers
